fix(og): reject an image value that is not a URL instead of shadowing the cover
`usableImageUrl` rejected blank text and nothing else, but the value is free
text an author types and `getImagePath` hands anything that is not `ipfs://`
straight to the `<img>`. So "hello" counted as an image, and at the front of the
chain it shadowed a cover that works.

The codebase already keeps this check — `isRenderableImageSrc`, whose comment
names the same hazard — so the chain reuses it rather than growing a second idea
of what a URL is.

It could not simply be imported: `utils.ts` imports `Entities`, so reaching back
the other way closes a cycle. It moves to a leaf module and `utils.ts`
re-exports it, leaving existing callers untouched.

The check belongs in the chain rather than in the OG route, because a bad value
has to fall through to the cover. Validating at the route would send it to the
default card instead, which is the worse half of the bug it fixes.
