+++ apps/web/core/utils/utils.ts
@@ -560,18 +560,9 @@ export const getImagePathAtLevel = (value: string, level: number) => {
 // Primary gateway (Filebase). For single-shot callers with no runtime fallback.
 export const getImagePath = (value: string) => getImagePathAtLevel(value, 0);
 
-// Image values are free-text entity properties, so an author can type anything.
-// next/image throws on a src that is neither root-relative nor an absolute URL,
-// which kills the whole page that rendered it.
-export const isRenderableImageSrc = (src: string) => {
-  if (src.startsWith('/')) return true;
-  try {
-    new URL(src);
-    return true;
-  } catch {
-    return false;
-  }
-};
+// Re-exported from its own leaf module so the share-image chain can use it without closing an
+// import cycle through this file's `Entities` import. Existing callers are unaffected.
+export { isRenderableImageSrc } from './image-src';
 
 export const getVideoHash = getImageHash;
 export const getVideoPathAtLevel = getImagePathAtLevel;
